Restructuration : Elements communs dans nouvelle app 'common' - fix panels manquants dans wagtail
