+++ apps/agenda/templates/agenda/includes/event_card.html
@@ -17,7 +17,7 @@
             <h3 class="text-xl font-bold mb-2 group-hover:text-[var(--primary-color)] transition-colors">
                 {{ event.title }}
             </h3>
-            <div class="text-gray-600 line-clamp-3">{{ event.description|richtext }}</div>
+            <div class="text-gray-600 line-clamp-3">{{ event.subtitle|default:event.content|richtext }}</div>
         </div>
         <a href="{{ event.url }}" class="btn-outline mt-2">
             En savoir plus
