+++ apps/common/templatetags/navigation_tags.py
@@ -1,5 +1,4 @@
 from django import template
-
 from wagtail.models import Site
 
 register = template.Library()
